refactor(unifi_network_rules): Reorganize rule utility functions into dedicated helpers module
- Move rule-related utility functions from `utils.py` to new `helpers/rule.py`
- Create separate `helpers` and `utils` packages for better code organization
- Maintain existing functionality for rule ID, name, and enabled state retrieval
- Update import statements in switch and UDM API modules to use new helper module
