+++ custom_components/unifi_network_rules/helpers/rule.py
@@ -1,4 +1,4 @@
-"""Utility functions for UniFi Network Rules."""
+"""Helper functions for UniFi Network rule handling."""
 from typing import Any
 
 from aiounifi.models.traffic_route import TrafficRoute
